JB23AT2-32-Get Hotel-room - change all connection for id and room number to be Long

+++ src/main/java/com/aacademy/moonlight/controller/hotel/RoomController.java
@@ -21,12 +21,12 @@ public class RoomController {
     RoomService roomService;
 
     @GetMapping("/find-room-by-id/{id}")
-    public ResponseEntity<RoomResponse> findRoomById (@Valid @PathVariable Integer id) {
+    public ResponseEntity<RoomResponse> findRoomById (@Valid @PathVariable Long id) {
      return ResponseEntity.status(HttpStatus.FOUND).body(roomService.getRoomById(id));
     }
 
     @GetMapping("/find-room-by-room-number/{roomNumber}")
-    public ResponseEntity<List<RoomResponse>> findRoomByRoomNumber(@Valid @PathVariable Integer roomNumber){
+    public ResponseEntity<List<RoomResponse>> findRoomByRoomNumber(@Valid @PathVariable Long roomNumber){
         List<RoomResponse>list = roomService.findByRoomNumber(roomNumber);
         return ResponseEntity.status(HttpStatus.FOUND).body(list);
     }
