+++ src/main/java/com/aacademy/moonlight/converter/hotel/RoomConverter.java
@@ -19,7 +19,7 @@ public class RoomConverter {
     private final RoomRepository repository;
     private final RoomFacilityRepository roomFacilityRepository;
 
-    public Room createRoom (Integer roomId, RoomRequest request){
+    public Room createRoom (Long roomId, RoomRequest request){
 
         /*
         List<Long> facilityIds = request.getFacilityIds();
